requirements.txt
# This file is used to install the required packages for the project.
reportlab
streamlit
openai
python-dotenv
pillow
pdf2image
opencv-python
PyMuPDF
XlsxWriter